dev: fix some sonarqube warnings
