+++ src/Playground.Console/Program.cs
@@ -8,6 +8,7 @@ using Playground.Common;
 using Playground.Common.Groups;
 using Playground.Common.Groups.TypealizR;
 using Playground.Common.NoCodeGen;
+using Playground.Console;
 
 const string arthur = "Arthur";
 const string chewbacca = "Chewbacca";
@@ -81,14 +82,6 @@ Console.WriteLine(
     without_Params_In_MethodNames.Goodbye(arthur)
 );
 
-#pragma warning restore CA1812 // Type 'Program' can be sealed because it has no subtypes in its containing assembly and is not externally visible
-
-
-
-
-
-
-
 
 
 var g = typealizedGroups;
@@ -100,7 +93,6 @@ static void SomeMethod(IStringLocalizer<Ressources> L)
 
 SomeMethod(g.Localizer);
 
-
 Console.WriteLine(
     g.Some.Deeply.Nested.Thing.Called.After.A.Monster.It
 );
@@ -110,24 +102,11 @@ Console.WriteLine(
 );
 
 
-
-
+//analyzer samples
 var localizer = provider.GetRequiredService<IStringLocalizer>();
 
-var a = localizer.Bar();
-var b = localizer.Bar("bar");
-var c = localizer.Foo("foo");
-var d = localizer.FooBar("fooBar");
-var bs = localizer.Bars("Arthur", 42);
-
-public static class IStringLocalizerExtensions
-{
-    public static LocalizedString Bar(this IStringLocalizer that) => that["Bar"];
-    public static LocalizedString Bar(this IStringLocalizer that, string value) => that["Bar", value];
-
-    public static LocalizedString Foo(this IStringLocalizer that, string value) => that["Foo {bar}", value];
-    public static LocalizedString FooBar(this IStringLocalizer that, string value) => that["FooBar {0}", value];
-
-    public static LocalizedString Bars(this IStringLocalizer that, string v1, int v2) => that["Bar {0} {1}", v1, v2];
-
-}
+Console.WriteLine(localizer.Bar());
+Console.WriteLine(localizer.Bar("bar"));
+Console.WriteLine(localizer.Foo("foo"));
+Console.WriteLine(localizer.FooBar("fooBar"));
+Console.WriteLine(localizer.Bars("Arthur", 42));
