+++ src/TypealizeR.Analyzers.CodeFixes/ICodeFixer.cs
@@ -10,7 +10,7 @@ internal interface ICodeFixer
     Task<Document> CreateChangedDocumentAsync(Document source, CancellationToken cancellationToken);
 }
 
-internal interface ICodeFixer<T> : ICodeFixer where T : SyntaxNode
+internal interface ICodeFixer<out T> : ICodeFixer where T : SyntaxNode
 {
     T Root { get; }
 }
